JAMES-2341 Move Role at API level
