+++ server/protocols/jmap/src/main/java/org/apache/james/jmap/json/RoleMixIn.java
@@ -16,22 +16,12 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
+package org.apache.james.jmap.json;
 
-package org.apache.james.jmap.utils;
+import com.fasterxml.jackson.annotation.JsonValue;
 
-import java.util.stream.Stream;
+public abstract class RoleMixIn {
 
-import org.apache.james.jmap.exceptions.MailboxRoleNotFoundException;
-import org.apache.james.mailbox.MailboxSession;
-import org.apache.james.mailbox.MessageManager;
-import org.apache.james.mailbox.Role;
-import org.apache.james.mailbox.exception.MailboxException;
-
-public interface SystemMailboxesProvider {
-    Stream<MessageManager> getMailboxByRole(Role aRole, MailboxSession session) throws MailboxException;
-
-    default MessageManager findMailbox(Role role, MailboxSession session) throws MailboxException {
-        return getMailboxByRole(role, session).findAny()
-            .orElseThrow(() -> new MailboxRoleNotFoundException(role));
-    }
+    @JsonValue
+    abstract String serialize();
 }
